refactor: 내비게이터, 메인 광고배너 렌더 함수를 각 컴포넌트와 index로 분리
