+++ public/js/index.js
@@ -1,7 +1,20 @@
+import { getToday } from "./util/utils.js";
 import { dowTopDefault } from "./views/blocks/dow-top.js";
-import { renderingDowNavDefault } from "./views/render/nav.js";
-import { renderingMainAd } from "./views/render/main-ad-banner.js";
+import { renderingMainAd } from "./views/components/main-ad-banner.js";
 import { handleCategory } from "./controllers/handle-cateogry.js";
+
+export const renderingNav = (selectedNav, selectedElement) => {
+  document.querySelectorAll(`.${selectedNav} li`).forEach((e) => {
+    e.id = "";
+  });
+  selectedElement.id = "selected";
+};
+
+const renderingDowNavDefault = () => {
+  document.querySelectorAll(".dow__nav .nav__list .nav__item").forEach((e) => {
+    if (e.dataset.value === getToday()) return renderingNav("dow__nav", e);
+  });
+};
 
 const renderingDefault = () => {
   renderingMainAd("홈");
